Move funder reports retrieval to funder home
