+++ src/scenes/AcceleratorRouter/FundingEntities/DeleteFundingEntityModal.tsx
@@ -6,7 +6,7 @@ import DialogBox from 'src/components/common/DialogBox/DialogBox';
 import Button from 'src/components/common/button/Button';
 import { APP_PATHS } from 'src/constants';
 import { useSyncNavigate } from 'src/hooks/useSyncNavigate';
-import FundingEntityService from 'src/services/FundingEntityService';
+import FundingEntityService from 'src/services/funder/FundingEntityService';
 import strings from 'src/strings';
 import { FundingEntity } from 'src/types/FundingEntity';
 import useSnackbar from 'src/utils/useSnackbar';
