+++ src/scenes/AcceleratorRouter/FundingEntities/InviteView.tsx
@@ -10,8 +10,8 @@ import PageForm from 'src/components/common/PageForm';
 import { EMAIL_REGEX } from 'src/constants';
 import useNavigateTo from 'src/hooks/useNavigateTo';
 import { useFundingEntity } from 'src/providers';
-import { requestFundingEntityInviteFunder } from 'src/redux/features/funder/fundingEntitiesAsyncThunks';
-import { inviteFunderRequest } from 'src/redux/features/funder/fundingEntitiesSelectors';
+import { requestFundingEntityInviteFunder } from 'src/redux/features/funder/entities/fundingEntitiesAsyncThunks';
+import { inviteFunderRequest } from 'src/redux/features/funder/entities/fundingEntitiesSelectors';
 import { useAppDispatch, useAppSelector } from 'src/redux/store';
 import strings from 'src/strings';
 import { Funder } from 'src/types/FundingEntity';
