+++ src/scenes/AcceleratorRouter/FundingEntities/ListView.tsx
@@ -8,8 +8,8 @@ import { FilterConfig } from 'src/components/common/SearchFiltersWrapperV2';
 import Button from 'src/components/common/button/Button';
 import useNavigateTo from 'src/hooks/useNavigateTo';
 import { useLocalization, useUser } from 'src/providers';
-import { requestFundingEntities } from 'src/redux/features/funder/fundingEntitiesAsyncThunks';
-import { selectFundingEntitiesRequest } from 'src/redux/features/funder/fundingEntitiesSelectors';
+import { requestFundingEntities } from 'src/redux/features/funder/entities/fundingEntitiesAsyncThunks';
+import { selectFundingEntitiesRequest } from 'src/redux/features/funder/entities/fundingEntitiesSelectors';
 import { useAppDispatch, useAppSelector } from 'src/redux/store';
 import strings from 'src/strings';
 import { FundingEntity } from 'src/types/FundingEntity';
